refactor: combine Rysk migration stages
